Update README.md
In the project so far, Kaju App Icon is successfully designated as the application icon by using Image Asset option. In all of the user interfaces, linear layout has been used. Splash screen is completed. Also, opening screen which the activity that takes the user's phone number. The algorithm takes the user's phone number and then compare with the phone numbers that are registered already in the firebase authentication. If the user is already registered to the system, program opens another screen and wants user to enter a 6 digit verification code. These activities and user interfaces are completed. In addition to this, main menu is almost completed.